Add custom controls for VideoScreen
Create the required video controls, including: restart video, fast backward, fast forward, and go to end of video.
Create the required styles to array controls in a fashion similar to the native player.
Ensure that RTL is working for controls.
Add animations to hide controls.
Add resume functionality when required.